face_detection_function/requirements.txt
google-cloud-storage
google-cloud-vision
google-cloud-firestore
